feat(contact): 新增通讯录功能
- 添加联系人数据库和相关 DAO
- 实现通讯录界面和逻辑
- 新增好友申请和验证功能
- 优化数据库版本和迁移策略
- 引入数据存储管理器

+++ app/src/main/java/com/example/youxin/network/model/response/UserinfoResp.kt
@@ -9,5 +9,5 @@ data class Info(
     val id: String,
     val nickname: String,
     val phone: String,
-    val sex: Int
+    val sex: Byte
 )
